+++ structurizr-core/src/com/structurizr/view/MetadataSymbols.java
@@ -1,0 +1,15 @@
+package com.structurizr.view;
+
+/**
+ * The type of symbols to use when rendering metadata.
+ */
+public enum MetadataSymbols {
+
+    SquareBrackets,
+    RoundBrackets,
+    CurlyBrackets,
+    AngledBrackets,
+    DoubleAngledBrackets,
+    None
+
+}